+++ app/app.js
@@ -29,7 +29,7 @@ const root = require('./src/apis/root');
 const review = require('./src/apis/review');
 const home = require('./src/apis/home');
 const schedule = require('./src/apis/schedule');
-const clubList = require('./src/apis/club');
+const clubList = require('./src/apis/club-list');
 const application = require('./src/apis/application');
 const clubBoard = require('./src/apis/club-board');
 const adminOption = require('./src/apis/admin-option');
