+++ app/src/apis/club-list/club.ctrl.js
@@ -1,12 +1,12 @@
 'use strict';
 
-const ClubStorage = require('../../models/services/Club/ClubStorage');
+const ClubListStorage = require('../../models/services/ClubList/ClubListStorage');
 const getApiInfo = require('../../models/utils/getApiInfo');
 const processCtrl = require('../../models/utils/processCtrl');
 
 const process = {
   readClubList: async (req, res) => {
-    const response = await ClubStorage.readClubList();
+    const response = await ClubListStorage.readClubList();
     const apiInfo = getApiInfo('GET', response, req);
 
     return processCtrl(res, apiInfo);
